Redid all the task/project junk.
Tasks are working pretty well.
Projects have been sort of implemented, but
you can't add project tasks yet.
I added some static files (logo, favicon)
I'm in the middle of fixing a few things...
like the logo in base.html and a dropdown button for projects
but this is likely where I'm stopping for the day